chore: reorganize project structure by moving navigation files into a dedicated folder
